requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
allure-pytest==2.9.45
allure-python-commons==2.9.45; python_version >= '3.5'
assertpy==1.1
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cerberus==1.3.4
certifi==2022.6.15; python_full_version >= '3.6.0'
charset-normalizer==2.1.0; python_full_version >= '3.6.0'
idna==3.3; python_version >= '3.5'
importlib-metadata==4.12.0; python_version < '3.8'
iniconfig==1.1.1
packaging==21.3; python_full_version >= '3.6.0'
pluggy==1.0.0; python_full_version >= '3.6.0'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest==7.1.2
python-dotenv==0.20.0
requests==2.28.1
setuptools==62.6.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version >= '3.7'
typing-extensions==4.3.0; python_version < '3.8'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.8.0; python_version >= '3.7'
